Update install.md
After carefully following all of the instructions, none of the tutorials or suggested commands could locate the docker executable. This was due to a missing entry in the PATH environmental variable. Uninstalling and following the instructions in this change request resolves this issue. A search for my issues turns up many suggestions all of which can be avoided if this hint is included here. First impressions are important, many people will walk away if following the existing instruction don't work.